Fixes #8, adds individual project pages
Themes can now specify a project.php and a projects.php file.
When the user navigates to /project.php the content of
themes/default/projects.php is shown, ideally containing a list of all the
projects.
When the user navigates to /project.php?id=1 the content of
themes/default/project.php is shown, which should contain code to display
a single project.